refactor: moved annotation to subpackage
